Update readme and move unit tests to focus on RepositoryBase
