+++ CorgiShop.Tests/Application/Requests/Products/DeleteProductsCommandHandlerTests.cs
@@ -4,16 +4,17 @@ using MediatR;
 using CorgiShop.Application.Features.Products.Commands.DeleteProduct;
 using CorgiShop.Domain.Abstractions;
 using CorgiShop.Domain.Model;
+using CorgiShop.Domain.Features.Products;
 
 namespace CorgiShop.Tests.Application.Requests.Products;
 
 public class DeleteProductsCommandHandlerTests
 {
-    private Mock<ICommandRepository<Product>> _mockedRepository;
+    private Mock<IProductRepository> _mockedRepository;
 
     public DeleteProductsCommandHandlerTests()
     {
-        _mockedRepository = new Mock<ICommandRepository<Product>>();
+        _mockedRepository = new Mock<IProductRepository>();
         _mockedRepository.Setup(r => r.Delete(0));
     }
